More code style fixes
