+++ src/hooks/useOrderRoute.ts
@@ -1,0 +1,14 @@
+import React from "react";
+import { OrderRouteContext } from "@/contexts/OrderRouteContext";
+
+/**
+ * Custom hook to use the OrderRoute context
+ * This hook provides access to order routing functionality
+ */
+export const useOrderRoute = () => {
+  const context = React.useContext(OrderRouteContext);
+  if (context === undefined) {
+    throw new Error("useOrderRoute must be used within an OrderRouteProvider");
+  }
+  return context;
+};
